requirements.txt
pdfplumber
PyYAML
fastapi
python-multipart
uvicorn
